Events are now displayed. Moved create event logic out of the CalendarView class into a static method in a separate class.
I moved the create event stuff out of CalendarView because that is not
CalendarView's responsibility. CalendarView displays a calendar with
selectable days - that is it. Sticking with MVC - creating an event is a
controller action, the CalendarView is a view. We don't want to mix
them.

I changed event back to using LocalTime objects for start and end times
- Date represents much more then just a time on the clock and that was
causing issues when displaying the time.

Note that currently you must enter 2 digit hours. 2:00 will fail, but
02:00 will work.